Add handling for OAuth2 code exchange and redirect user to account page upon callback
This commit adds a new file, `route.ts`, located in `src/app/auth/callback/` to handle the exchange of an OAuth2 code for a session token via the `createRouteHandlerClient` function from `@supabase/auth-helpers-nextjs`. The function `GET` extracts the code from the request URL and calls `supabase.auth.exchangeCodeForSession` with it - if the code exists. Finally, it returns a `NextResponse` object that redirects the user to the account page.