Change target compiler option to es5 and set strict mode to true.
In detail, this commit changes the target compiler option to es5, enabling compatibility with older browsers. Additionally, the commit sets the `strict` option to `true` in `compilerOptions` of `tsconfig.json` file, indicating that TypeScript is to apply additional checks. It also corrects the position of three lines involving a plugin name, example paths, and files for compilation. These changes stabilize the codebase and avoid possible compatibility issues.