Create a form for user login with input validations
The commit adds an AuthForm in the src/components directory which is a form that takes username and password as input. The form has implemented input validation with the help of zod library and useForm hook from react-hook-form. Form components such as Button, Form, FormField, FormLabel, FormControl and Input are used from the UI directory.